jobs/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile jobs/requirements.in -o jobs/requirements.txt
aiohappyeyeballs==2.4.0
    # via aiohttp
aiohttp==3.10.5
    # via
    #   -r jobs/requirements.in
    #   llama-index-core
aiolimiter==1.1.0
    # via -r jobs/requirements.in
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
asttokens==2.4.1
    # via stack-data
asyncpg==0.29.0
    # via
    #   -r jobs/requirements.in
    #   llama-index-vector-stores-postgres
attrs==24.2.0
    # via aiohttp
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via nltk
dataclasses-json==0.6.7
    # via llama-index-core
decorator==5.1.1
    # via
    #   ipython
    #   retry
deprecated==1.2.14
    # via llama-index-core
dirtyjson==1.0.8
    # via llama-index-core
distro==1.9.0
    # via openai
executing==2.1.0
    # via stack-data
filelock==3.16.1
    # via
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.9.0
    # via
    #   huggingface-hub
    #   llama-index-core
    #   torch
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.2
    # via
    #   langsmith
    #   llama-index-core
    #   openai
huggingface-hub==0.25.1
    # via
    #   tokenizers
    #   transformers
idna==3.8
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
ipython==8.27.0
    # via pyvis
jedi==0.19.1
    # via ipython
jinja2==3.1.4
    # via
    #   pyvis
    #   torch
jiter==0.5.0
    # via openai
joblib==1.4.2
    # via nltk
jsonpatch==1.33
    # via langchain-core
jsonpickle==3.3.0
    # via pyvis
jsonpointer==3.0.0
    # via jsonpatch
langchain-core==0.2.38
    # via langchain-openai
langchain-openai==0.1.23
    # via -r jobs/requirements.in
langsmith==0.1.117
    # via langchain-core
llama-index-core==0.11.14
    # via
    #   -r jobs/requirements.in
    #   llama-index-embeddings-openai
    #   llama-index-graph-stores-neo4j
    #   llama-index-llms-openai
    #   llama-index-llms-openai-like
    #   llama-index-llms-openrouter
    #   llama-index-storage-docstore-postgres
    #   llama-index-storage-kvstore-postgres
    #   llama-index-vector-stores-postgres
llama-index-embeddings-openai==0.2.5
    # via -r jobs/requirements.in
llama-index-graph-stores-neo4j==0.3.2
    # via -r jobs/requirements.in
llama-index-llms-openai==0.2.9
    # via llama-index-llms-openai-like
llama-index-llms-openai-like==0.2.0
    # via llama-index-llms-openrouter
llama-index-llms-openrouter==0.2.0
    # via -r jobs/requirements.in
llama-index-storage-docstore-postgres==0.2.0
    # via -r jobs/requirements.in
llama-index-storage-kvstore-postgres==0.2.0
    # via llama-index-storage-docstore-postgres
llama-index-vector-stores-postgres==0.2.5
    # via -r jobs/requirements.in
markupsafe==2.1.5
    # via jinja2
marshmallow==3.22.0
    # via dataclasses-json
matplotlib-inline==0.1.7
    # via ipython
mpmath==1.3.0
    # via sympy
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
neo4j==5.25.0
    # via llama-index-graph-stores-neo4j
nest-asyncio==1.6.0
    # via llama-index-core
networkx==3.3
    # via
    #   llama-index-core
    #   pyvis
    #   torch
nltk==3.9.1
    # via llama-index-core
numpy==1.26.4
    # via
    #   llama-index-core
    #   pgvector
    #   transformers
openai==1.44.1
    # via
    #   langchain-openai
    #   llama-index-embeddings-openai
    #   llama-index-llms-openai
orjson==3.10.7
    # via langsmith
packaging==24.1
    # via
    #   huggingface-hub
    #   langchain-core
    #   marshmallow
    #   transformers
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pgvector==0.2.5
    # via llama-index-vector-stores-postgres
pillow==10.4.0
    # via llama-index-core
prompt-toolkit==3.0.48
    # via ipython
psycopg2==2.9.9
    # via -r jobs/requirements.in
psycopg2-binary==2.9.9
    # via llama-index-vector-stores-postgres
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
py==1.11.0
    # via retry
pydantic==2.9.1
    # via
    #   -r jobs/requirements.in
    #   langchain-core
    #   langsmith
    #   llama-index-core
    #   openai
pydantic-core==2.23.3
    # via pydantic
pygments==2.18.0
    # via ipython
pytz==2024.2
    # via neo4j
pyvis==0.3.2
    # via -r jobs/requirements.in
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   langchain-core
    #   llama-index-core
    #   transformers
redis==5.0.8
    # via -r jobs/requirements.in
regex==2024.7.24
    # via
    #   nltk
    #   tiktoken
    #   transformers
requests==2.32.2
    # via
    #   -r jobs/requirements.in
    #   huggingface-hub
    #   langsmith
    #   llama-index-core
    #   tiktoken
    #   transformers
retry==0.9.2
    # via -r jobs/requirements.in
safetensors==0.4.5
    # via transformers
setuptools==75.1.0
    # via torch
six==1.16.0
    # via asttokens
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
sqlalchemy==2.0.32
    # via
    #   -r jobs/requirements.in
    #   llama-index-core
    #   llama-index-vector-stores-postgres
stack-data==0.6.3
    # via ipython
sympy==1.13.3
    # via torch
tenacity==8.5.0
    # via
    #   langchain-core
    #   llama-index-core
tiktoken==0.7.0
    # via
    #   langchain-openai
    #   llama-index-core
tokenizers==0.20.0
    # via transformers
torch==2.4.1
    # via -r jobs/requirements.in
tqdm==4.66.5
    # via
    #   huggingface-hub
    #   llama-index-core
    #   nltk
    #   openai
    #   transformers
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
transformers==4.45.1
    # via llama-index-llms-openai-like
typing-extensions==4.12.2
    # via
    #   huggingface-hub
    #   langchain-core
    #   llama-index-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   torch
    #   typing-inspect
typing-inspect==0.9.0
    # via
    #   dataclasses-json
    #   llama-index-core
urllib3==2.2.2
    # via requests
wcwidth==0.2.13
    # via prompt-toolkit
wrapt==1.16.0
    # via
    #   deprecated
    #   llama-index-core
yarl==1.11.1
    # via aiohttp
